refactored class structure
